Revert "Move types inside src"
This reverts commit f7444f4e5a24ea182527a16269f0d83f9e328c78.
