+++ fixtures/configurations/basic.ts
@@ -1,4 +1,4 @@
-import {AbstractOperationType, ConfigurationType} from '../../src/types/domain';
+import {AbstractOperationType, ConfigurationType} from '../../types/domain';
 
 export function minimalConfiguration(): ConfigurationType {
   {
